change type

+++ src/main/java/com/esprit/alternance/kaddem/repositories/EquipeRepository.java
@@ -4,7 +4,7 @@ import com.esprit.alternance.kaddem.entities.Equipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface EquipeRepository extends JpaRepository<Equipe, Integer> {
+public interface EquipeRepository extends JpaRepository<Equipe, Long> {
